requirements.txt
PyYAML>=6.0
nats-py>=2.6.0
